Fix Issue #12, reported by Christian.Strempfer - change Path.getTotalLength() to return a double instead of an int - change Path.getPointAtLength() to accept a double instead of an int

+++ trunk/src/com/hydro4ge/raphaelgwt/client/RaphaelJS.java
@@ -378,10 +378,10 @@ class RaphaelJS extends JavaScriptObject {
    */
   protected static class Path extends Element {
     protected Path() {}
-    public final native int getTotalLength() /*-{
+    public final native double getTotalLength() /*-{
       return this.getTotalLength();
     }-*/;
-    public final native Point getPointAtLength(int length) /*-{
+    public final native Point getPointAtLength(double length) /*-{
       return this.getPointAtLength(length);
     }-*/;
     public final native String getSubpath(int from, int to) /*-{
